src/app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
databricks-sdk>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
fpdf2>=2.7.0
mlflow>=2.15.0
